pyproject.toml
[project]
name = "chess-study-assistant"
version = "1.0.0"
description = "Chess study agent"
authors = [
    {name = "santiago-chica",email = "santiago.chica1997@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-chess (>=1.999,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

